+++ geotrek/trekking/forms.py
@@ -4,6 +4,7 @@ from django.utils.translation import gettext as _
 from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.forms.models import inlineformset_factory
+
 from django import forms
 
 from crispy_forms.helper import FormHelper